Added testbuild tag to trigger Actions -pipeline
The 'testbuild' tag can be used to trigger the GitHub Actions workflow. It can
be used to build and run on any update, and the main point is that releases
are only created on pushes to the master branch.

Usage: pushing a new branch with an update to a package
$ git push -u origin <new branch name here> testbuild