GitHub Actions: retry connecting to the VM to avoid failure
Sometimes the VM boots in the Actions runner, but SSH fails with:

    kex_exchange_identification: read: Connection reset by peer
    Error: Process completed with exit code 255.

This is probably caused by some delay when OpenSSH is generating its
host keys. In such a situation, all we have to do is to wait for OpenSSH
to finish its initialization before connecting.

Try connecting 5 times before the "real" one which will make the test
fail if it fails.